[CDF-21406] Remove common folder function code (#480)
* refactor: remove all common folder code

* refactor: remove common code and update functions

* tests: Regen

* refactor: update run local function without common folder

* build: changelog

* refactor: consistent naming

* refactor; remove ref to commont_func_code

* refactor; done

---------
