+++ cdf-tk-dev.py
@@ -47,7 +47,6 @@ if not Path(REPO_ROOT / "cognite_toolkit" / "config.local.yaml").exists():
             project="<my-project>",
             build_type="dev",
             selected_modules_and_packages=["cdf_demo_infield", "cdf_oid_example_data"],
-            common_function_code="./common_function_code",
         )
     ).load_defaults(REPO_ROOT / "cognite_toolkit")
     (REPO_ROOT / "cognite_toolkit" / "config.local.yaml").write_text(config_init.dump_yaml_with_comments())
