+++ cognite_toolkit/_cdf_tk/templates/_migration.yaml
@@ -1,3 +1,13 @@
+- version: 0.2.0a4
+  cognite_modules: {}
+  resources: {}
+  tool:
+    - title: 'cognite-toolkit no longer supports common_function_code'
+      steps:
+        - For each function that is using code from the `common_function_code`,
+          copy the relevant code for that function from the `common_function_code`
+          into the function folder.
+  cognite_modules_hash: ""
 - version: 0.2.0a3
   cognite_modules: {}
   resources: {}
@@ -12,7 +22,7 @@
     - title: "In all `config.<env>.yaml` the section 'modules' is renamed to 'variables"
       steps:
         - For each `config.<env>.yaml` file, rename the section `modules` to `variables`.
-  cognite_modules_hash: ""
+  cognite_modules_hash: "6c19979abb10571a513af8e530f7807a54687fd292b0572ff9c5fa41c82df640"
 - version: 0.2.0a2
   cognite_modules: {}
   resources: {}
