+++ tests/tests_unit/test_approval_modules_snapshots/cdf_data_pipeline_files_valhall.yaml
@@ -76,7 +76,7 @@ Function:
   externalId: fn_context_files_oid_fileshare_annotation
   functionPath: handler.py
   metadata:
-    cdf-toolkit-function-hash: 8092971a985b3c9b5bc7566903f4cf67d039a12a819d46b0b69ccb1fbac73af7
+    cdf-toolkit-function-hash: ef732c41127c8ee3bc266c9d7734279d8ea8293912a5243060284d3c37daa7f3
     version: 0.0.1
   name: context:files:oid:fileshare:annotation
   owner: Anonymous
