+++ tests/tests_unit/test_approval_modules_snapshots/cdf_data_pipeline_timeseries_valhall.yaml
@@ -134,7 +134,7 @@ Function:
   externalId: fn_context_timeseries_oid_opcua_asset
   functionPath: ./handler.py
   metadata:
-    cdf-toolkit-function-hash: 706e21bb09813aa765fa0f10cd183e9fbfa712ffc49cf11e98032c0a18be8773
+    cdf-toolkit-function-hash: 751c0304548641f6407bfaecbd52b3f1f589a87a17e9799c28a6f77d8cd410b6
     version: 0.0.1
   name: context:timeseries:oid:opcua:asset
   owner: Anonymous
