+++ tests/tests_unit/test_cdf_tk/run_data/cognite_modules/examples/run_local_functions/functions/fn_test2/requirements.txt
@@ -1,1 +1,1 @@
-cognite-sdk>=7.13.8
+cognite-sdk>=7.37.0
