+++ tests/tests_unit/test_cdf_tk/test_build_files.py
@@ -6,7 +6,6 @@ import typer
 from pytest import MonkeyPatch
 
 from cognite_toolkit._cdf import build
-from tests.constants import REPO_ROOT
 from tests.tests_unit.utils import mock_read_yaml_file
 
 
@@ -19,7 +18,6 @@ def mock_environments_yaml_file(module_name: str, monkeypatch: MonkeyPatch) -> N
                     "project": "pytest-project",
                     "type": "dev",
                     "selected_modules_and_packages": [module_name],
-                    "common_function_code": Path(REPO_ROOT / "cognite_toolkit/common_function_code").as_posix(),
                 }
             }
         },
