+++ tests_migrations/constants.py
@@ -11,6 +11,7 @@ SUPPORTED_TOOLKIT_VERSIONS = [
     "0.1.2",
     "0.2.0a1",
     "0.2.0a2",
+    "0.2.0a3",
 ]
